requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
passlib[bcrypt]
bcrypt
PyJWT
python-dotenv
pydantic
numpy
resampy
webrtcvad
faster-whisper
openai
backoff
pyyaml
httpx
